modified:   packages/client/index.html 	modified:   packages/client/package.json 	renamed:    packages/client/src/App.css -> packages/client/src/components/app/App.css 	renamed:    packages/client/src/App.test.tsx -> packages/client/src/components/app/App.test.tsx 	renamed:    packages/client/src/App.tsx -> packages/client/src/components/app/App.tsx 	modified:   packages/client/src/index.css 	modified:   packages/client/src/main.tsx 	new file:   packages/client/src/redux/hooks.ts 	new file:   packages/client/src/redux/store.ts 	modified:   yarn.lock
